# openapi.yaml
openapi: 3.0.3
info:
  title: SoleSigner API
  description: Minimalistic serverless voting system with cryptographic receipts
  version: 0.1.0
  contact:
    name: SoleSigner Team
    url: https://github.com/your-org/solesigner
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new admin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '201':
          description: Admin registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
        '400':
          description: Validation error
        '409':
          description: Username already exists
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 50
        password:
          type: string
          minLength: 6
          maxLength: 100
...